refactor: Moving Port Canvas components and renaming Props
