+++ src/components/shared/SimpleCard.tsx
@@ -1,10 +1,10 @@
-interface Props {
+interface SimpleCardProps {
   title: string;
   subtitle: string;
   className?: string;
 }
 
-function SimpleCard({title, subtitle, className}: Props) {
+function SimpleCard({title, subtitle, className}: SimpleCardProps) {
   return (
     <div className={`simple-card ${className}`}>
       <h3 className="mb-4 text-xl font-bold">{title}</h3>
